Create utility classes and enhance existing ones in MBMigration
This update introduces new utility classes (TimeUtility, FoldersUtility) in the MBMigration module to handle time formatting and directory creation. It also adds new functionalities to others like array sorting in ArrayManipulator and page manipulation in MigrationPlatform. These changes provide a more organized structure for data handling in the migration process.
