+++ lib/MBMigration/Layer/MB/MBProjectDataCollector.php
@@ -62,7 +62,6 @@ class MBProjectDataCollector
         ];
     }
 
-
     /**
      * @throws Exception
      */
